+++ zsdk/zia/__init__.py
@@ -16,6 +16,7 @@ from .rule_labels import rule_labels
 from .sandbox_report import sandbox_report_file, sandbox_report_quota
 from .sandbox_settings import sandbox_settings
 from .sandbox_submission import sandbox_submission
+from .security_policy_settings import allowlist, denylist, blacklist
 from .user_management import departments, groups, users, auditors
 from .firewall_policies import (
     firewall_filtering_rules,
@@ -244,3 +245,15 @@ class zia:
     @property
     def sandbox_submission(self) -> sandbox_submission:
         return sandbox_submission(session=self._session, base_url=self._base_url)
+
+    @property
+    def allowlist(self) -> allowlist:
+        return allowlist(session=self._session, base_url=self._base_url)
+
+    @property
+    def denylist(self) -> denylist:
+        return denylist(session=self._session, base_url=self._base_url)
+
+    @property
+    def blacklist(self) -> blacklist:
+        return blacklist(session=self._session, base_url=self._base_url)